some tests, not finished
